requirements.txt
# Web Framework
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20
jinja2==3.1.6

# Session & Auth
itsdangerous==2.2.0

# SSE Streaming
sse-starlette==2.2.1

# AI & LangGraph
langchain==0.3.14
langchain-community==0.3.14
langchain-core==0.3.47
langchain-openai==0.3.0
langchain-google-genai==2.1.1
langchain-text-splitters==0.3.7
langgraph==0.2.64
langgraph-checkpoint==2.0.21
langgraph-sdk==0.1.58
langsmith==0.2.11
openai==1.68.2

# Search
tiktoken==0.9.0

# Database & ORM
sqlalchemy==2.0.38
alembic==1.14.1
pymysql==1.1.1

# Auth & Passwords
passlib[bcrypt]==1.7.4
pyjwt==2.10.1

# AWS Storage
boto3==1.36.26

# PDF Generation
reportlab==4.3.1
Markdown==3.7
beautifulsoup4==4.13.3

# Utilities
python-dotenv==1.0.1
pydantic==2.10.6
pydantic-settings==2.8.1
pillow==11.1.0
aiohttp==3.11.14
httpx==0.28.1

# HTTP & Async
anyio==4.9.0
httpcore==1.0.7
